Se ha modificado el tipo del tamaño de los archivos de la clase FileInfo pasan de ser flotantes representando el tamaño en MegaBytes a enteros para representar el tamaño en Bytes, por lo tanto tambien se cambio la manera de representar el limite del tamaño donado en el fichero de configuracion, pasando de ser flotante a ser entero. El motivo del cambio es ganar exactitud a la hora de permitir si un fichero se puede guardar en un peer o no.

+++ src/main/java/es/udc/fic/tic/nautilus/config/NautilusConfig.java
@@ -8,13 +8,13 @@ import es.udc.fic.tic.nautilus.util.ModelConstanst.LANGUAGE;
 public class NautilusConfig {
 
 	private boolean serverAvailable;
-	private Float limitSpace;
+	private int limitSpace;
 	private ModelConstanst.LANGUAGE language;
 	private List<ServerInfo> serverPreferences;
 	private String storageFolder;
 	
 	
-	public NautilusConfig(boolean serverAvailable, Float limitSpace, LANGUAGE language,
+	public NautilusConfig(boolean serverAvailable, int limitSpace, LANGUAGE language,
 			List<ServerInfo> serverPreferences, String storageFolder) {
 		super();
 		this.serverAvailable = serverAvailable;
@@ -33,11 +33,11 @@ public class NautilusConfig {
 		this.serverAvailable = serverAvailable;
 	}
 	
-	public Float getLimitSpace() {
+	public int getLimitSpace() {
 		return limitSpace;
 	}
 	
-	public void setLimitSpace(Float limitSpace) {
+	public void setLimitSpace(int limitSpace) {
 		this.limitSpace = limitSpace;
 	}
